+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @cat-factory/provider-s3
 
+## 0.2.357
+
+### Patch Changes
+
+- Updated dependencies [1fbd83c]
+- Updated dependencies [00228c6]
+  - @cat-factory/kernel@0.281.2
+
 ## 0.2.356
 
 ### Patch Changes